Moved constants to the correct place...
